+++ ios/Classes/SwiftRemindersPlugin.swift
@@ -3,6 +3,7 @@ import Flutter
 public class SwiftRemindersPlugin: NSObject, FlutterPlugin {
 
   let reminders = Reminders()
+  let calendars = Events()
 
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(name: "reminders", binaryMessenger: registrar.messenger())
@@ -49,7 +50,7 @@ public class SwiftRemindersPlugin: NSObject, FlutterPlugin {
           }
         }
 
-    case "deleteReminder":
+      case "deleteReminder":
       if let args = call.arguments as? [String: String] {
         if let id = args["id"] {
           self.reminders.deleteReminder(id) { (error) in
@@ -58,7 +59,28 @@ public class SwiftRemindersPlugin: NSObject, FlutterPlugin {
         }
       }
 
-      default:
+      case "requestAccess":
+        result (self.calendars.requestAccess())
+
+      case "hasEventsAccess":
+        result(self.calendars.hasEventsAccess())
+
+      case "getDefaultCalendar":
+        result(self.calendars.getDefaultCalendar())
+
+      case "getAllCalendars":
+        result(self.calendars.getAllCalendars())
+
+      case "getEvents":
+        if let args = call.arguments as? [String: String?] {
+          if let id = args["id"] {
+              self.calendars.getEvents(id) { (event) in
+                result(event)
+            }
+          }
+        }
+
+    default:
         result(FlutterMethodNotImplemented)
     }
   }